refactor run helpers
