+++ wandb/latest-run
@@ -1,1 +1,1 @@
-run-20230309_130822-tgm99uu8
+run-20230309_140629-n4cwhw3n
