+++ wandb/run-20230309_140541-n4cwhw3n/files/requirements.txt
@@ -1,0 +1,215 @@
+absl-py==1.4.0
+affine==2.4.0
+aiohttp==3.8.4
+aiosignal==1.3.1
+appdirs==1.4.4
+asciitree==0.3.3
+asttokens==2.2.1
+async-timeout==4.0.2
+attrs==22.2.0
+backcall==0.2.0
+backports.functools-lru-cache==1.6.4
+beautifulsoup4==4.11.2
+black==23.1.0
+blinker==1.5
+bokeh==2.4.3
+bottleneck==1.3.6
+branca==0.6.0
+brotlipy==0.7.0
+cachetools==5.3.0
+certifi==2022.12.7
+cffi==1.15.1
+cfgrib==0.9.10.3
+cfgv==3.3.1
+cftime==1.6.2
+charset-normalizer==2.1.1
+click-plugins==1.1.1
+click==8.1.3
+cligj==0.7.2
+cloudpickle==2.2.1
+colorama==0.4.6
+comm==0.1.2
+contourpy==1.0.7
+coverage==7.2.0
+cryptography==39.0.1
+cycler==0.11.0
+cytoolz==0.12.0
+dask-glm==0.2.0
+dask-ml==2022.5.27
+dask==2023.2.0
+debugpy==1.6.6
+decorator==5.1.1
+distlib==0.3.6
+distributed==2023.2.0
+docker-pycreds==0.4.0
+docopt==0.6.2
+eccodes==1.5.1
+einops==0.6.0
+entrypoints==0.4
+exceptiongroup==1.1.0
+executing==1.2.0
+fasteners==0.17.3
+fastjsonschema==2.16.2
+filelock==3.9.0
+findlibs==0.0.2
+finetuning-scheduler==0.2.3
+fiona==1.9.1
+flake8==6.0.0
+flox==0.6.8
+folium==0.14.0
+fonttools==4.38.0
+frozenlist==1.3.3
+fsspec==2023.1.0
+gcsfs==2023.1.0
+gdal==3.6.2
+geopandas==0.12.2
+gitdb==4.0.10
+gitpython==3.1.31
+google-api-core==2.11.0
+google-auth-oauthlib==0.4.6
+google-auth==2.16.1
+google-cloud-core==2.3.2
+google-cloud-storage==2.7.0
+google-crc32c==1.1.2
+google-resumable-media==2.4.1
+googleapis-common-protos==1.57.1
+grpcio==1.52.1
+h5netcdf==1.1.0
+h5py==3.8.0
+heapdict==1.0.1
+identify==2.5.18
+idna==3.4
+importlib-metadata==6.0.0
+importlib-resources==5.12.0
+iniconfig==2.0.0
+ipykernel==6.21.2
+ipython==8.10.0
+ipytree==0.2.2
+ipywidgets==8.0.4
+jedi==0.18.2
+jinja2==3.1.2
+joblib==1.2.0
+jsonschema==4.17.3
+jupyter-client==8.0.3
+jupyter-core==5.2.0
+jupyterlab-widgets==3.0.5
+kiwisolver==1.4.4
+lightning-utilities==0.6.0.post0
+llvmlite==0.39.1
+locket==1.0.0
+lz4==4.3.2
+mapclassify==2.5.0
+markdown==3.4.1
+markupsafe==2.1.2
+matplotlib-inline==0.1.6
+matplotlib==3.7.0
+mccabe==0.7.0
+msgpack==1.0.4
+multidict==6.0.4
+multipledispatch==0.6.0
+munch==2.5.0
+munkres==1.1.4
+mypy-extensions==1.0.0
+nbformat==5.7.3
+nc-time-axis==1.4.1
+nest-asyncio==1.5.6
+netcdf4==1.6.2
+networkx==3.0
+nodeenv==1.7.0
+numba==0.56.4
+numbagg==0.2.2
+numcodecs==0.11.0
+numpy-groupies==0.9.20
+numpy==1.23.5
+oauthlib==3.2.2
+packaging==23.0
+pandas==1.5.3
+parso==0.8.3
+partd==1.3.0
+pathspec==0.11.0
+pathtools==0.1.2
+pexpect==4.8.0
+pickleshare==0.7.5
+pillow==9.4.0
+pip==23.0.1
+pkgutil-resolve-name==1.3.10
+platformdirs==3.0.0
+pluggy==1.0.0
+pooch==1.6.0
+pre-commit==3.0.4
+prompt-toolkit==3.0.36
+protobuf==4.21.12
+psutil==5.9.4
+ptyprocess==0.7.0
+pure-eval==0.2.2
+pyasn1-modules==0.2.7
+pyasn1==0.4.8
+pycodestyle==2.10.0
+pycparser==2.21
+pydap==3.2.2
+pydeprecate==0.3.2
+pyflakes==3.0.1
+pygments==2.14.0
+pyjwt==2.6.0
+pyopenssl==23.0.0
+pyparsing==3.0.9
+pyproj==3.4.1
+pyrsistent==0.19.3
+pysocks==1.7.1
+pytest==7.2.1
+python-dateutil==2.8.2
+pytorch-lightning==1.7.7
+pytz==2022.7.1
+pyu2f==0.1.5
+pyyaml==6.0
+pyzmq==25.0.0
+rasterio==1.3.4
+requests-oauthlib==1.3.1
+requests==2.28.2
+rioxarray==0.13.4
+rsa==4.9
+rtree==1.0.1
+scikit-learn==1.2.1
+scipy==1.10.1
+seaborn==0.12.2
+sentry-sdk==1.16.0
+setproctitle==1.3.2
+setuptools==67.4.0
+shapely==2.0.1
+six==1.16.0
+smmap==3.0.5
+snuggs==1.4.7
+sortedcontainers==2.4.0
+soupsieve==2.4
+stack-data==0.6.2
+tblib==1.7.0
+tensorboard-data-server==0.7.0
+tensorboard-plugin-wit==1.8.1
+tensorboard==2.12.0
+threadpoolctl==3.1.0
+tomli==2.0.1
+toolz==0.12.0
+torch==1.13.1
+torchmetrics==0.11.1
+torchvision==0.14.1
+tornado==6.2
+tqdm==4.64.1
+traitlets==5.9.0
+typing-extensions==4.4.0
+ukkonen==1.0.1
+unicodedata2==15.0.0
+urllib3==1.26.14
+virtualenv==20.19.0
+wandb==0.13.10
+wcwidth==0.2.6
+webob==1.8.7
+werkzeug==2.2.3
+wheel==0.38.4
+widgetsnbextension==4.0.5
+xarray==2023.2.0
+xbatcher==0.3.0
+xyzservices==2023.2.0
+yarl==1.8.2
+zarr==2.13.6
+zict==2.2.0
+zipp==3.14.0
