+++ wandb/run-20230309_140541-n4cwhw3n/files/wandb-metadata.json
@@ -1,0 +1,64 @@
+{
+    "os": "Linux-4.19.0-23-cloud-amd64-x86_64-with-glibc2.28",
+    "python": "3.9.16",
+    "heartbeatAt": "2023-03-09T14:05:41.772112",
+    "startedAt": "2023-03-09T14:05:41.415990",
+    "docker": null,
+    "cuda": "11.0.228",
+    "args": [
+        "--model_class=vit.simpleVIT",
+        "--limit_val_batches=5",
+        "--limit_train_batches=10",
+        "--max_epochs=5",
+        "--log_every_n_steps=2",
+        "--ft_schedule=hyperiap/litmodels/LitClassifier_ft_schedule_final.yaml"
+    ],
+    "state": "running",
+    "program": "/home/glennmoncrieff/hyper-iap/run_all.py",
+    "codePath": "run_all.py",
+    "git": {
+        "remote": "https://github.com/GMoncrieff/hyper-iap.git",
+        "commit": "3dbd0fb555668b9de79c165d672cab4d51b930e4"
+    },
+    "email": "glennwithtwons@gmail.com",
+    "root": "/home/glennmoncrieff/hyper-iap",
+    "host": "rs-dl",
+    "username": "glennmoncrieff",
+    "executable": "/home/glennmoncrieff/mambaforge/envs/rs-hyper/bin/python",
+    "cpu_count": 2,
+    "cpu_count_logical": 4,
+    "cpu_freq": {
+        "current": 2249.998,
+        "min": 0.0,
+        "max": 0.0
+    },
+    "cpu_freq_per_core": [
+        {
+            "current": 2249.998,
+            "min": 0.0,
+            "max": 0.0
+        },
+        {
+            "current": 2249.998,
+            "min": 0.0,
+            "max": 0.0
+        },
+        {
+            "current": 2249.998,
+            "min": 0.0,
+            "max": 0.0
+        },
+        {
+            "current": 2249.998,
+            "min": 0.0,
+            "max": 0.0
+        }
+    ],
+    "disk": {
+        "total": 98.18611526489258,
+        "used": 42.13631057739258
+    },
+    "memory": {
+        "total": 31.41571807861328
+    }
+}
